% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_associations.R
\name{compute_initial_associations}
\alias{compute_initial_associations}
\title{Deploy associations across datasets}
\usage{
compute_initial_associations(
  bound_data,
  primary_variable,
  constant_adjusters = NULL,
  model_type = "glm",
  proportion_cutoff = 1,
  vibrate = TRUE,
  family = gaussian(),
  ids = NULL,
  strata = NULL,
  weights = NULL,
  nest = NULL
)
}
\arguments{
\item{bound_data}{merged independent an dependent data for all datasets}

\item{primary_variable}{The column name from the independent_variables tibble containing the key variable you want to associate with disease in your first round of modeling (prior to vibration). For example, if you are interested fundamentally identifying how well age can predict height, you would make this value a string referring to whatever column in said dataframe refers to "age."}

\item{constant_adjusters}{A character vector (or just one string) of column names corresponding to column names in your dataset to include in every vibration. (default = NULL)}

\item{model_type}{Specifies regression type -- "glm", "survey", or "negative_binomial". Survey regression will require additional parameters (at leaset weight, nest, strata, and ids). Any model family (e.g. gaussian()), or any other parameter can be passed as an additional argument to this function.}

\item{proportion_cutoff}{Float between 0 and 1. Filter out dependent features that are this proportion of zeros or more (default = 1, so no filtering done.)}

\item{vibrate}{TRUE/FALSE -- run vibrations (default=TRUE)}

\item{family}{GLM family (default = gaussian()). For help see help(glm) or help(family).}

\item{ids}{Name of column in dataframe specifying cluster ids from largest level to smallest level. Only relevant for survey data. (Default = NULL).}

\item{strata}{Name of column in dataframe with strata. Relevant for survey data. (Default = NULL).}

\item{weights}{Name of column containing sampling weights.}

\item{nest}{If TRUE, relabel cluster ids to enforce nesting within strata.}
}
\description{
Top-level function to run all associations for all datasets.
}
\keyword{association}
\keyword{initial}
\keyword{regression,}
