% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_metaanalysis.R
\name{compute_metaanalysis}
\alias{compute_metaanalysis}
\title{Run meta-analysis}
\usage{
compute_metaanalysis(df)
}
\arguments{
\item{df}{Association output.}
}
\description{
Run meta-analysis for each feature
}
\keyword{meta-analysis}
