% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanesToyDataset.R
\docType{data}
\name{dependent_toy_dataset1_nhanes}
\alias{dependent_toy_dataset1_nhanes}
\title{Toy NHANES datasets (not for use in actual analysis) for exploring VoE}
\format{
dataframe
}
\source{
NHANES, <https://www.cdc.gov/nchs/nhanes/index.html>
}
\usage{
data(nhanesToyDataset)
}
\description{
Data from NHANES 1994-2000, arbitrarily split into 2 datasets 
to demonstrate how the VoE pipeline works
}
\keyword{datasets}
