% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_linear_output.R
\name{filter_unnest_feature_vib}
\alias{filter_unnest_feature_vib}
\title{Unnest vibration data}
\usage{
filter_unnest_feature_vib(vib_df)
}
\arguments{
\item{vib_df}{VoE dataframe with columns for each adjuster (output of get_adjuster_expanded_vibrations).}
}
\description{
Unnest regression output for vibrations.
}
\keyword{analysis}
\keyword{voe}
