% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_linear_output.R
\name{find_confounders_linear}
\alias{find_confounders_linear}
\title{Find confounders}
\usage{
find_confounders_linear(voe_list_for_reg)
}
\arguments{
\item{voe_list_for_reg}{A dataframe of expanded VoE output (output of filter_unnest_feature_vib)}
}
\description{
Model confounding from vibration analysis.
}
\keyword{analysis}
\keyword{voe}
