% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_linear_output.R
\name{get_adjuster_expanded_vibrations}
\alias{get_adjuster_expanded_vibrations}
\title{Expand vibration data}
\usage{
get_adjuster_expanded_vibrations(voe_df, adjusters, constant_adjusters)
}
\arguments{
\item{voe_df}{Raw vibration output, the first entry in the output of compute_vibrations.}

\item{adjusters}{A dataframe, each column corresponding to the adjusters used in each dataset for vibrations. This is the second entry in the compute_vibrations output.}

\item{constant_adjusters}{A character vector (or just one string) corresponding to column names in your dataset to include in every vibration. (default = NULL)}
}
\description{
Add column to vibration output for each adjuster (indicating presence or absence in a model).
}
\keyword{analysis}
\keyword{voe}
