% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_metaanalysis.R
\name{get_converged_metadfs}
\alias{get_converged_metadfs}
\title{Filter-meta analysis}
\usage{
get_converged_metadfs(meta_df, dataset_num)
}
\arguments{
\item{meta_df}{Meta-analysis output.}

\item{dataset_num}{Number of datasets.}
}
\description{
Remove failed meta-analyses.
}
\keyword{meta-analysis}
