% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_metaanalysis.R
\name{get_summary_stats}
\alias{get_summary_stats}
\title{Extract meta-analysis summary statistics}
\usage{
get_summary_stats(input_meta_df, dataset_num)
}
\arguments{
\item{input_meta_df}{Meta-analysis output.}

\item{dataset_num}{Number of datasets.}
}
\description{
Remove failed meta-analyses.
}
\keyword{meta-analysis}
