% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independent_variable_analysis.R
\name{ind_var_analysis}
\alias{ind_var_analysis}
\title{Analysis of Independent Variables}
\usage{
ind_var_analysis(independent_variables, output_location = getwd())
}
\arguments{
\item{independent_variables}{A tibble containing the information for your independent variables (e.g. age, sex). Each column should correspond to a different variable (e.g. age), with the first column containing the sample names matching those in the column anmes of the dependent_variables tibble.}

\item{output_location}{The path where you want to save all the ggplots and metadata}
}
\description{
This function will run the full pipeline
}
\keyword{independent}
\keyword{variable}
