% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independent_variable_analysis.R
\name{makeHistogram}
\alias{makeHistogram}
\title{Making a Histogram}
\usage{
makeHistogram(table, columnName, pathToNewFolder)
}
\arguments{
\item{table}{A tibble containing the information for your independent variables (e.g. age, sex). Each column should correspond to a different variable (e.g. age), with the first column containing the sample names matching those in the column anmes of the dependent_variables tibble.}

\item{columnName}{The name of the column of which data will correspond to the x axis}

\item{pathToNewFolder}{The path - a new folder - where the ggplot bar graph will be saved to}
}
\description{
This function will create a Bar Graph for discrete data
}
\keyword{histogram}
