% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_metaanalysis.R
\name{meta_analysis_command}
\alias{meta_analysis_command}
\title{Meta-analysis function}
\usage{
meta_analysis_command(te, sete, studlab, name)
}
\arguments{
\item{te}{estimates}

\item{sete}{standard errors}

\item{studlab}{study labels}

\item{name}{study labels}
}
\description{
metagen function
}
\keyword{meta-analysis}
