% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_pre_pipeline.R
\name{pre_pipeline_data_check}
\alias{pre_pipeline_data_check}
\title{Pre-flight checks}
\usage{
pre_pipeline_data_check(
  dependent_variables,
  independent_variables,
  primary_variable,
  fdr_method,
  fdr_cutoff,
  max_vibration_num,
  max_vars_in_model,
  proportion_cutoff,
  meta_analysis,
  model_type,
  family,
  ids,
  strata,
  weights,
  nest
)
}
\arguments{
\item{dependent_variables}{A tibble containing the information for your dependent variables (e.g. bacteria relative abundance, age). The columns should correspond to different variables, and the rows should correspond to different units, such as individuals (e.g. individual1, individual2, etc). If passing multiple datasets, pass a named list of the same length and in the same order as the independent_variables parameter. If running from the command line, pass the path (or comma separated paths) to .rds files containing the data, one per dataset.}

\item{independent_variables}{A tibble containing the information for your independent variables (e.g. bacteria relative abundance, age). The columns should correspond to different variables, and the rows should correspond to different units,  (e.g. individual1, individual2, etc). If passing multiple datasets, pass a named list of the same length and in the same order as the dependent_variables parameter. If running from the command line, pass the path (or comma separated paths) to .rds files containing the data, one per dataset.}

\item{primary_variable}{The column name from the independent_variables tibble containing the key variable you want to associate with disease in your first round of modeling (prior to vibration). For example, if you are interested fundamentally identifying how well age can predict height, you would make this value a string referring to whatever column in said dataframe refers to "age."}

\item{fdr_method}{Your choice of method for adjusting p-values. Options are BY (default), BH, or bonferroni. Adjustment is computed for all initial, single variable, associations across all dependent features.}

\item{fdr_cutoff}{Cutoff for an FDR significant association. All features with adjusted p-values initially under this value will be selected for vibrations. (default = 0.05). Setting a stringent FDR cutoff is mostly relevant when you are using a large number of dependent variables (eg >50) and want to filter those with weak initial associations.}

\item{max_vibration_num}{Maximum number of vibrations allowed for a single dependent variable. Setting this will also reduce runtime by reducing the number of models fit. (default = 10,000)}

\item{max_vars_in_model}{Maximum number of variables allowed in a single fit in vibrations. In case an individual has many hundreds of metadata features, this prevents models from being fit with excessive numbers of variables. Setting this parameter will improve runtime for large datasets. (default=NULL)}

\item{proportion_cutoff}{Float between 0 and 1. Filter out dependent features that are this proportion of zeros or more (default = 1, so no filtering will be done.)}

\item{meta_analysis}{TRUE/FALSE -- indicates if computing meta-analysis across multiple datasets. Set to TRUE by default if the pipeline detects multiple datasets. Setting this variable to TRUE but providing one dataset will throw an error.}

\item{model_type}{Specifies regression type -- "glm", "survey", or "negative_binomial". Survey regression will require additional parameters (at least weight, nest, strata, and ids). Any model family (e.g. gaussian()), or any other parameter can be passed as the family argument to this function.}

\item{family}{GLM family (default = gaussian()). For help see help(glm) or help(family).}

\item{ids}{Name of column in dataframe specifying cluster ids from largest level to smallest level. Only relevant for survey data. (Default = NULL).}

\item{strata}{Name of column in dataframe with strata. Relevant for survey data. (Default = NULL).}

\item{weights}{Name of column containing sampling weights.}

\item{nest}{If TRUE, relabel cluster ids to enforce nesting within strata.}
}
\description{
Check datasets and print pre-run statistics prior to deployment.
}
\keyword{pipeline}
