% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_linear_output.R
\name{summarize_vibration_data_by_feature}
\alias{summarize_vibration_data_by_feature}
\title{summarize_vibration_data_by_feature}
\usage{
summarize_vibration_data_by_feature(df)
}
\arguments{
\item{df}{A dataframe of expanded VoE output (output of filter_unnest_feature_vib)}
}
\description{
Summarize output of vibrations for each dependent feature of interest.
}
\keyword{analysis}
\keyword{voe}
