% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_linear_output.R
\name{analyze_voe_data}
\alias{analyze_voe_data}
\title{Analyze VoE data}
\usage{
analyze_voe_data(vibration_output, confounder_analysis)
}
\arguments{
\item{vibration_output}{Output list from the compute vibrations function.}

\item{confounder_analysis}{TRUE/FALSE -- run confounder analysis (default = TRUE).}
}
\description{
Post-process vibration output.
}
\keyword{analysis}
\keyword{voe}
