% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_associations.R
\name{compute_initial_associations}
\alias{compute_initial_associations}
\title{Deploy associations across datasets}
\usage{
compute_initial_associations(
  bound_data,
  primary_variable,
  model_type,
  proportion_cutoff,
  vibrate,
  regression_weights
)
}
\arguments{
\item{bound_data}{merged independent an dependent data for all datasets}

\item{primary_variable}{The column name from the independent_variables tibble containing the key variable you want to associate with disease in your first round of modeling (prior to vibration). For example, if you are interested fundamentally identifying how well age can predict height, you would make this value a string referring to whatever column in said dataframe refers to "age."}

\item{model_type}{Model family (e.g. gaussian, binomial, etc). Will determine if you are doing classification or regression. See GLM families for more information. (default="gaussian")}

\item{proportion_cutoff}{Float between 0 and 1. Filter out dependent features that are this proportion of zeros or more (default = 1, so no filtering done.)}

\item{vibrate}{TRUE/FALSE -- run vibrations (default=TRUE)}

\item{regression_weights}{Column in independent variable dataset(s) corresponding to weights  for linear regression input (default = NULL).}
}
\description{
Top-level function to run all associations for all datasets.
}
\keyword{assocatiation}
\keyword{initial}
\keyword{regression,}
