% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_vibrations.R
\name{compute_vibrations}
\alias{compute_vibrations}
\title{Vibrations}
\usage{
compute_vibrations(
  bound_data,
  primary_variable,
  model_type,
  features_of_interest,
  max_vibration_num,
  proportion_cutoff,
  cores,
  max_vars_in_model,
  family,
  ids,
  strata,
  weights,
  nest
)
}
\arguments{
\item{bound_data}{Dataframe of tibbles. All independent and depenendent dataframes for all datasets.}

\item{primary_variable}{The column name from the independent_variables tibble containing the key variable you want to associate with disease in your first round of modeling (prior to vibration). For example, if you are interested fundamentally identifying how well age can predict height, you would make this value a string referring to whatever column in said dataframe refers to "age."}

\item{model_type}{Specifies regression type -- "glm", "survey", or "negative_binomial". Survey regression will require additional parameters (at least weight, nest, strata, and ids). Any model family (e.g. gaussian()), or any other parameter can be passed as the family argument to this function.}

\item{features_of_interest}{Feature to vibrate over.}

\item{max_vibration_num}{Maximum number of vibrations allowed for a single dependent variable. Setting this will also reduce runtime by reducing the number of models fit. (default = 10,000)}

\item{proportion_cutoff}{Float between 0 and 1. Filter out dependent features that are this proportion of zeros or more (default = 1, so no filtering will be done.)}

\item{cores}{Number of threads.}

\item{max_vars_in_model}{Maximum number of variables allowed in in a model.}

\item{family}{GLM family (default = gaussian()). For help see help(glm) or help(family).}

\item{ids}{Name of column in dataframe specifying cluster ids from largest level to smallest level. Only relevant for survey data. (Default = NULL).]}

\item{strata}{Name of column in dataframe with strata. Only relevant for survey data. (Default = NULL).}

\item{weights}{Name of column containing sampling weights.}

\item{nest}{If TRUE, relabel cluster ids to enforce nesting within strata}
}
\description{
Run vibrations for all features for all datasets
}
\keyword{association}
\keyword{initial}
\keyword{regression,}
