% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_vibrations.R
\name{dataset_vibration}
\alias{dataset_vibration}
\title{Vibration for dataset}
\usage{
dataset_vibration(
  subframe,
  primary_variable,
  model_type,
  features_of_interest,
  max_vibration_num,
  proportion_cutoff,
  regression_weights,
  cores,
  max_vars_in_model
)
}
\arguments{
\item{subframe}{List of length 2. Dataframes containing a single datasets independent and dependent data.}

\item{primary_variable}{The column name from the independent_variables tibble containing the key variable you want to associate with disease in your first round of modeling (prior to vibration). For example, if you are interested fundamentally identifying how well age can predict height, you would make this value a string referring to whatever column in said dataframe refers to "age."}

\item{model_type}{Model family (e.g. gaussian, binomial, etc). Will determine if you are doing classification or regression. See GLM families for more information. (default="gaussian")}

\item{features_of_interest}{Feature to vibrate over.}

\item{max_vibration_num}{Maximum number of vibrations (default=50000).}

\item{proportion_cutoff}{Float between 0 and 1. Filter out dependent features that are this proportion of zeros or more (default = 1, so no filtering done).}

\item{regression_weights}{Column in independent variable dataset(s) corresponding to weights  for linear regression input (default = NULL).}

\item{cores}{Number of threads.}

\item{max_vars_in_model}{Maximum number of variables allowed in in a model.}
}
\description{
Run vibrations for all features in a dataset
}
\keyword{assocatiation}
\keyword{initial}
\keyword{regression,}
