% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_pre_pipeline.R
\name{pre_pipeline_data_check}
\alias{pre_pipeline_data_check}
\title{Pre-flight checks}
\usage{
pre_pipeline_data_check(
  dependent_variables,
  independent_variables,
  primary_variable,
  fdr_method,
  fdr_cutoff,
  max_vibration_num,
  max_vars_in_model,
  proportion_cutoff,
  meta_analysis,
  model_type
)
}
\arguments{
\item{dependent_variables}{A tibble containing the information for your dependent variables (e.g. bacteria relative abundance, age). The first column should be the rownames (e.g. gene1, gene2, gene3), and the columns should correspond to different samples (e.g. individual1, individual2, etc).}

\item{independent_variables}{A tibble containing the independent variables you will want to vibrate over. Each column should correspond to a different variable (e.g. age), with the first column containing the sample names matching those in the column names of the dependent_variables tibble.}

\item{primary_variable}{The column name from the independent_variables tibble containing the key variable you want to associate with disease in your first round of modeling (prior to vibration). For example, if you are interested fundamentally identifying how well age can predict height, you would make this value a string referring to whatever column in said dataframe refers to "age."}

\item{fdr_method}{Your choice of method for adjusting p-values. Options are BY (default), BH, or bonferroni.}

\item{fdr_cutoff}{Cutoff for an FDR significant association (default = 0.05).}

\item{max_vibration_num}{Maximum number of vibrations (default=50000).}

\item{max_vars_in_model}{Maximum number of variables allowed in a single fit (for vibrations). (default=NULL)}

\item{proportion_cutoff}{Float between 0 and 1. Filter out dependent features that are this proportion of zeros or more (default = 1, so no filtering done.)}

\item{meta_analysis}{TRUE/FALSE -- indicates if computing meta-analysis across multiple datasets.}

\item{model_type}{Model family (e.g. gaussian, binomial, etc). Will determine if you are doing classification or regression. See GLM families for more information. (default="gaussian")}
}
\description{
Check datasets and print pre-run statistics prior to deployment.
}
\keyword{pipeline}
