% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_associations.R
\name{regression}
\alias{regression}
\title{Run initial association}
\usage{
regression(
  j,
  independent_variables,
  dependent_variables,
  primary_variable,
  model_type,
  proportion_cutoff,
  regression_weights
)
}
\arguments{
\item{j}{feature name}

\item{independent_variables}{A tibble containing the independent variables you will want to vibrate over. Each column should correspond to a different variable (e.g. age), with the first column containing the sample names matching those in the column names of the dependent_variables tibble.}

\item{dependent_variables}{A tibble containing the information for your dependent variables (e.g. bacteria relative abundance, age). The first column should be the rownames (e.g. gene1, gene2, gene3), and the columns should correspond to different samples (e.g. individual1, individual2, etc).}

\item{primary_variable}{The column name from the independent_variables tibble containing the key variable you want to associate with disease in your first round of modeling (prior to vibration). For example, if you are interested fundamentally identifying how well age can predict height, you would make this value a string referring to whatever column in said dataframe refers to "age."}

\item{model_type}{Model family (e.g. gaussian, binomial, etc). Will determine if you are doing classification or regression. See GLM families for more information. (default="gaussian")}

\item{proportion_cutoff}{Float between 0 and 1. Filter out dependent features that are this proportion of zeros or more (default = 1, so no filtering done.)}

\item{regression_weights}{Column in independent variable dataset(s) corresponding to weights  for linear regression input (default = NULL).}
}
\description{
Run initial association for a single feature
}
\keyword{assocatiation}
\keyword{initial}
\keyword{regression,}
