% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_vibrations.R
\name{vibrate}
\alias{vibrate}
\title{Vibration for feature}
\usage{
vibrate(
  merged_data,
  variables_to_vibrate,
  max_vars_in_model,
  feature,
  primary_variable,
  model_type,
  regression_weights,
  max_vibration_num,
  dataset_id,
  proportion_cutoff
)
}
\arguments{
\item{merged_data}{Merged independent and dependent data.}

\item{variables_to_vibrate}{Variables over which you're going to vibrate.}

\item{max_vars_in_model}{Maximum number of variables allowed in in a model.}

\item{feature}{Feature over which to vibrate.}

\item{primary_variable}{The column name from the independent_variables tibble containing the key variable you want to associate with disease in your first round of modeling (prior to vibration). For example, if you are interested fundamentally identifying how well age can predict height, you would make this value a string referring to whatever column in said dataframe refers to "age."}

\item{model_type}{Model family (e.g. gaussian, binomial, etc). Will determine if you are doing classification or regression. See GLM families for more information. (default="gaussian")}

\item{regression_weights}{Column in independent variable dataset(s) corresponding to weights  for linear regression input (default = NULL).}

\item{max_vibration_num}{Maximum number of vibrations (default=50000).}

\item{dataset_id}{Identifier for dataset.}

\item{proportion_cutoff}{Float between 0 and 1. Filter out dependent features that are this proportion of zeros or more (default = 1, so no filtering done.)}
}
\description{
Run vibrations for a single feature
}
\keyword{assocatiation}
\keyword{initial}
\keyword{regression,}
